requirements.txt
# Core dependencies
opencv-python==4.8.1.78
moviepy==1.0.3
Pillow==10.1.0
numpy==1.24.3

# Audio processing
edge-tts==6.1.9
pydub==0.25.1

# Video processing
ffmpeg-python==0.2.0
imageio==2.31.5
imageio-ffmpeg==0.4.9

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
colorlog==6.7.0
tqdm==4.66.1

# API (optional)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pytest-cov==4.1.0